+++ CHANGELOG.md
@@ -7,6 +7,9 @@ FlatLaf Change Log
 
 - HTML: Fixed wrong rendering if HTML text contains `<style>` tag with
   attributes (e.g. `<style type='text/css'>`). (issue #905; regression in 3.5.1)
+- FlatLaf window decorations:
+  - Linux: Fixed continuous cursor toggling between resize and standard cursor
+    when resizing window. (issue #907)
 
 
 ## 3.5.2